events-python/requirements.txt
flask==3.0.3
redis==5.0.4
pika==1.3.2
python-dotenv==1.0.1
